requirements.txt
loguru>=0.5.3
tdameritrade
mysql.connector